+++ app/models/__init__.py
@@ -1,3 +1,3 @@
-from app.models.user import User, Role
-from app.models.task import Task, Priority, Status
-from app.models.notification import Notification
+from app.models.user_model import User, Role
+from app.models.task_model import Task, Priority, Status
+from app.models.notification_model import Notification
