+++ app/routers/notifications.py
@@ -6,8 +6,8 @@ from sqlalchemy import select, update
 from sqlalchemy.orm import Session
 
 from app.core.database import get_db
-from app.models.notification import Notification
-from app.models.user import User
+from app.models.notification_model import Notification
+from app.models.user_model import User
 from app.routers.users import get_current_user
 from app.schemas.notification import NotificationOut  # <-- import BEFORE usage
