+++ app/routers/tasks.py
@@ -4,8 +4,8 @@ from fastapi import APIRouter, Depends, HTTPException, Query, status
 from sqlalchemy.orm import Session
 from sqlalchemy import select, and_, or_
 from app.core.database import get_db
-from app.models.task import Task, Priority, Status
-from app.models.user import User, Role
+from app.models.task_model import Task, Priority, Status
+from app.models.user_model import User, Role
 from app.schemas.task import TaskCreate, TaskUpdate, TaskOut
 from app.routers.users import get_current_user
