+++ app/routers/users.py
@@ -3,7 +3,7 @@ from fastapi.security import OAuth2PasswordBearer
 from sqlalchemy.orm import Session
 from app.core.database import get_db
 from app.core.security import decode_token
-from app.models.user import User, Role
+from app.models.user_model import User, Role
 from app.schemas.user import UserOut
 
 router = APIRouter(prefix="/users", tags=["users"])
