+++ tests/test_pagination.py
@@ -9,9 +9,9 @@ sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from app.core.database import Base
 from app.core.security import hash_password
-from app.models.user import User, Role
-from app.models.task import Task
-from app.models.notification import Notification
+from app.models.user_model import User, Role
+from app.models.task_model import Task
+from app.models.notification_model import Notification
 from app.routers.tasks import list_tasks
 from app.routers.notifications import list_my_notifications
